// src/core/domain/enums/papel-atributo-analitico.ts
/**
 * Papel Semântico do Atributo Analítico (Subunidade 3.6A)
 * Função exercida pela coluna/atributo no modelo analítico
 */
export enum PapelAtributoAnalitico {
  CHAVE_PRIMARIA = 'CHAVE_PRIMARIA',
  CHAVE_ESTRANGEIRA = 'CHAVE_ESTRANGEIRA',
  METRICA_BASE = 'METRICA_BASE',
  DIMENSAO_TEMPO = 'DIMENSAO_TEMPO',
  ATRIBUTO_DESCRITIVO = 'ATRIBUTO_DESCRITIVO',
}