// src/core/domain/enums/cardinalidade-relacionamento.ts
/**
 * Cardinalidade do Relacionamento Analítico (Subunidade 3.6A)
 * Proporção de linhas entre entidades conectadas
 */
export enum CardinalidadeRelacionamento {
  UM_PARA_UM = 'UM_PARA_UM',
  UM_PARA_MUITOS = 'UM_PARA_MUITOS',
  MUITOS_PARA_UM = 'MUITOS_PARA_UM',
  MUITOS_PARA_MUITOS = 'MUITOS_PARA_MUITOS',
}