// src/core/domain/enums/unidade-medida-metrica.ts
/**
 * Unidade de Medida da Métrica Analítica (Subunidade 3.6A)
 * Representação semântica do formato e unidade do indicador
 */
export enum UnidadeMedidaMetrica {
  MOEDA = 'MOEDA',
  PERCENTUAL = 'PERCENTUAL',
  QUANTIDADE = 'QUANTIDADE',
  TEMPO = 'TEMPO',
  TEXTO = 'TEXTO',
  INDICE = 'INDICE',
}